// apps/web/src/api/generated/model/advanceStageDtoToStage.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type AdvanceStageDtoToStage = typeof AdvanceStageDtoToStage[keyof typeof AdvanceStageDtoToStage];


export const AdvanceStageDtoToStage = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  hired: 'hired',
  rejected: 'rejected',
  withdrawn: 'withdrawn',
} as const;